DjangoProjectBase/requirements.txt
openai
python-dotenv
numpy
requests
matplotlib
django
pillow


